consolidate saver dependencies
  - let ezSaver() convenience routine require anything
    - use __dirname in makeReviver() to isolate from
      caller's directory
  - move io builders (fsReadAccess ,...) to saver.js
  - pass saver to run() rather than dbfile
